requirements.txt
requests
pandas
pyyaml
pycountry
